Move blob validation to EngineNewPayloadV3 class
