+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/engine/EngineNewPayloadV3.java
@@ -14,21 +14,27 @@
  */
 package org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.engine;
 
+import static org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.ExecutionEngineJsonRpcMethod.EngineStatus.INVALID;
 import static org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcError.INVALID_PARAMS;
 import static org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcError.UNSUPPORTED_FORK;
 
 import org.hyperledger.besu.consensus.merge.blockcreation.MergeMiningCoordinator;
+import org.hyperledger.besu.datatypes.VersionedHash;
 import org.hyperledger.besu.ethereum.ProtocolContext;
 import org.hyperledger.besu.ethereum.api.jsonrpc.RpcMethod;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.parameters.EnginePayloadParameter;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcErrorResponse;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcResponse;
+import org.hyperledger.besu.ethereum.core.Transaction;
 import org.hyperledger.besu.ethereum.eth.manager.EthPeers;
 import org.hyperledger.besu.ethereum.mainnet.ProtocolSchedule;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 import io.vertx.core.Vertx;
+import org.apache.tuweni.bytes.Bytes32;
 
 public class EngineNewPayloadV3 extends AbstractEngineNewPayload {
 
@@ -66,6 +72,58 @@ public class EngineNewPayloadV3 extends AbstractEngineNewPayload {
           || payloadParameter.getExcessDataGas() != null) {
         return Optional.of(new JsonRpcErrorResponse(reqId, UNSUPPORTED_FORK));
       }
+    }
+    return Optional.empty();
+  }
+
+  @Override
+  protected Optional<JsonRpcResponse> validateTransactions(
+      final EnginePayloadParameter blockParam,
+      final Object reqId,
+      final List<Transaction> transactions,
+      final Optional<List<String>> maybeVersionedHashParam) {
+
+    var blobTransactions =
+        transactions.stream().filter(transaction -> transaction.getType().supportsBlob()).toList();
+
+    return validateBlobTransactions(blockParam, reqId, blobTransactions, maybeVersionedHashParam);
+  }
+
+  private Optional<JsonRpcResponse> validateBlobTransactions(
+      final EnginePayloadParameter blockParam,
+      final Object reqId,
+      final List<Transaction> blobTransactions,
+      final Optional<List<String>> maybeVersionedHashParam) {
+
+    List<Bytes32> versionedHashesParam =
+        maybeVersionedHashParam
+            .map(strings -> strings.stream().map(Bytes32::fromHexStringStrict).toList())
+            .orElseGet(ArrayList::new);
+
+    final List<Bytes32> transactionVersionedHashes = new ArrayList<>();
+
+    for (Transaction transaction : blobTransactions) {
+      if (transaction.getType().supportsBlob()) {
+        var versionedHashes = transaction.getVersionedHashes();
+        if (versionedHashes.isEmpty()) {
+          return Optional.of(
+              respondWithInvalid(
+                  reqId, blockParam, null, INVALID, "There must be at least one blob"));
+        }
+        transactionVersionedHashes.addAll(
+            versionedHashes.get().stream().map(VersionedHash::toBytes).toList());
+      }
+    }
+
+    // check list contents
+    if (!versionedHashesParam.equals(transactionVersionedHashes)) {
+      return Optional.of(
+          respondWithInvalid(
+              reqId,
+              blockParam,
+              null,
+              INVALID,
+              "Versioned hashes from blob transactions do not match expected values"));
     }
     return Optional.empty();
   }
